requirements.txt
easyocr==1.7.2
keyboard==0.13.5
mss==10.1.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydirectinput==1.0.4
rapidfuzz==3.14.1
torch==2.5.0+cpu
torchvision==0.20.0+cpu